requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9

# Authentication
djangorestframework-simplejwt>=5.3.0

# File Storage
django-storages>=1.14.2
boto3>=1.34.0

# Background Jobs
celery>=5.3.4
redis>=5.0.1

# PDF Generation
reportlab>=4.0.7
qrcode>=7.4.2
Pillow>=10.2.0

# Utilities
python-dotenv>=1.0.0
pytz>=2024.1
django-filter>=23.5

# API Documentation
drf-spectacular>=0.27.0

# Security
django-ratelimit>=4.1.0

# Development
django-debug-toolbar>=4.2.0

# Monitoring (optional)
flower>=2.0.1

